+++ default_www/backend/modules/blog/engine/model.php
@@ -62,14 +62,14 @@ class BackendBlogModel
 		// init var
 		$warnings = array();
 
-		// blog rss title
+		// rss title
 		if(BackendModel::getModuleSetting('blog', 'rss_title_'. BL::getWorkingLanguage(), null) == '')
 		{
 			// add warning
 			$warnings[] = array('message' => sprintf(BL::err('RSSTitle', 'blog'), BackendModel::createURLForAction('settings', 'blog')));
 		}
 
-		// blog rss description
+		// rss description
 		if(BackendModel::getModuleSetting('blog', 'rss_description_'. BL::getWorkingLanguage(), null) == '')
 		{
 			// add warning
@@ -82,10 +82,10 @@ class BackendBlogModel
 
 
 	/**
-	 * Deletes one or more blogposts
+	 * Deletes one or more items
 	 *
 	 * @return	void
-	 * @param 	mixed $ids						The ids to delete.
+	 * @param 	mixed $ids		The ids to delete.
 	 */
 	public static function delete($ids)
 	{
@@ -95,7 +95,7 @@ class BackendBlogModel
 		// get db
 		$db = BackendModel::getDB(true);
 
-		// delete blogpost records
+		// delete records
 		$db->delete('blog_posts', 'id IN ('. implode(',', $ids) .') AND language = ?', array(BL::getWorkingLanguage()));
 		$db->delete('blog_comments', 'post_id IN ('. implode(',', $ids) .') AND language = ?', array(BL::getWorkingLanguage()));
 
@@ -116,7 +116,7 @@ class BackendBlogModel
 	 * Deletes a category
 	 *
 	 * @return	void
-	 * @param	int $id							The id of the category to delete.
+	 * @param	int $id		The id of the category to delete.
 	 */
 	public static function deleteCategory($id)
 	{
@@ -144,7 +144,7 @@ class BackendBlogModel
 	 * Deletes one or more comments
 	 *
 	 * @return	void
-	 * @param	array $ids						The id(s) of the comment(s) to delete.
+	 * @param	array $ids		The id(s) of the comment(s) to delete.
 	 */
 	public static function deleteComments($ids)
 	{
@@ -154,8 +154,8 @@ class BackendBlogModel
 		// get db
 		$db = BackendModel::getDB(true);
 
-		// get blogpost ids
-		$postIds = (array) $db->getColumn('SELECT i.post_id
+		// get ids
+		$itemIds = (array) $db->getColumn('SELECT i.post_id
 											FROM blog_comments AS i
 											WHERE i.id IN ('. implode(',', $ids) .') AND i.language = ?', array(BL::getWorkingLanguage()));
 
@@ -163,7 +163,7 @@ class BackendBlogModel
 		$db->delete('blog_comments', 'id IN ('. implode(',', $ids) .') AND language = ?', array(BL::getWorkingLanguage()));
 
 		// recalculate the comment count
-		if(!empty($postIds)) self::reCalculateCommentCount($postIds);
+		if(!empty($itemIds)) self::reCalculateCommentCount($itemIds);
 
 		// invalidate the cache for blog
 		BackendModel::invalidateFrontendCache('blog', BL::getWorkingLanguage());
@@ -180,8 +180,8 @@ class BackendBlogModel
 		// get db
 		$db = BackendModel::getDB(true);
 
-		// get blogpost ids
-		$postIds = (array) $db->getColumn('SELECT i.post_id
+		// get ids
+		$itemIds = (array) $db->getColumn('SELECT i.post_id
 											FROM blog_comments AS i
 											WHERE status = ? AND i.language = ?', array('spam', BL::getWorkingLanguage()));
 
@@ -189,7 +189,7 @@ class BackendBlogModel
 		$db->delete('blog_comments', 'status = ? AND language = ?', array('spam', BL::getWorkingLanguage()));
 
 		// recalculate the comment count
-		if(!empty($postIds)) self::reCalculateCommentCount($postIds);
+		if(!empty($itemIds)) self::reCalculateCommentCount($itemIds);
 
 		// invalidate the cache for blog
 		BackendModel::invalidateFrontendCache('blog', BL::getWorkingLanguage());
@@ -197,10 +197,10 @@ class BackendBlogModel
 
 
 	/**
-	 * Checks if a blogpost exists
+	 * Checks if an item exists
 	 *
 	 * @return	bool
-	 * @param	int $id							The id of the blogpost to check for existence.
+	 * @param	int $id		The id of the item to check for existence.
 	 */
 	public static function exists($id)
 	{
@@ -215,7 +215,7 @@ class BackendBlogModel
 	 * Checks if a category exists
 	 *
 	 * @return	int
-	 * @param	int $id							The id of the category to check for existence.
+	 * @param	int $id		The id of the category to check for existence.
 	 */
 	public static function existsCategory($id)
 	{
@@ -230,7 +230,7 @@ class BackendBlogModel
 	 * Checks if a comment exists
 	 *
 	 * @return	int
-	 * @param	int $id							The id of the comment to check for existence.
+	 * @param	int $id		The id of the comment to check for existence.
 	 */
 	public static function existsComment($id)
 	{
@@ -245,7 +245,7 @@ class BackendBlogModel
 	 * Get all data for a given id
 	 *
 	 * @return	array
-	 * @param	int $id							The Id of the blogpost to fetch?
+	 * @param	int $id		The Id of the item to fetch?
 	 */
 	public static function get($id)
 	{
@@ -253,8 +253,7 @@ class BackendBlogModel
 															m.url
 															FROM blog_posts AS i
 															INNER JOIN meta AS m ON m.id = i.meta_id
-															WHERE i.id = ? AND i.status = ? AND i.language = ?
-															LIMIT 1',
+															WHERE i.id = ? AND i.status = ? AND i.language = ?',
 															array((int) $id, 'active', BL::getWorkingLanguage()));
 	}
 
@@ -365,7 +364,7 @@ class BackendBlogModel
 	 * Get all data for a given id
 	 *
 	 * @return	array
-	 * @param	int $id							The id of the category to fetch.
+	 * @param	int $id		The id of the category to fetch.
 	 */
 	public static function getCategory($id)
 	{
@@ -401,7 +400,7 @@ class BackendBlogModel
 	 * Get all data for a given id
 	 *
 	 * @return	array
-	 * @param	int $id							The Id of the comment to fetch?
+	 * @param	int $id		The Id of the comment to fetch?
 	 */
 	public static function getComment($id)
 	{
@@ -420,7 +419,7 @@ class BackendBlogModel
 	 * Get multiple comments at once
 	 *
 	 * @return	array
-	 * @param	array $ids						The id(s) of the comment(s).
+	 * @param	array $ids		The id(s) of the comment(s).
 	 */
 	public static function getComments(array $ids)
 	{
@@ -449,8 +448,8 @@ class BackendBlogModel
 	 * Get the latest comments for a given type
 	 *
 	 * @return	array
-	 * @param	string $status					The status for the comments to retrieve.
-	 * @param	int[optional] $limit			The maximum number of items to retrieve.
+	 * @param	string $status			The status for the comments to retrieve.
+	 * @param	int[optional] $limit	The maximum number of items to retrieve.
 	 */
 	public static function getLatestComments($status, $limit = 10)
 	{
@@ -492,8 +491,8 @@ class BackendBlogModel
 	 * Get all data for a given revision
 	 *
 	 * @return	array
-	 * @param	int $id							The id of the blogpost.
-	 * @param	int $revisionId					The revision to get.
+	 * @param	int $id				The id of the item.
+	 * @param	int $revisionId		The revision to get.
 	 */
 	public static function getRevision($id, $revisionId)
 	{
@@ -510,9 +509,9 @@ class BackendBlogModel
 	 *
 	 * @return	string
 	 * @param	string $URL			The URL to base on.
-	 * @param	int[optional] $itemId		The id of the blogpost to ignore.
+	 * @param	int[optional] $id	The id of the item to ignore.
 	 */
-	public static function getURL($URL, $itemId = null)
+	public static function getURL($URL, $id = null)
 	{
 		// redefine URL
 		$URL = SpoonFilter::urlise((string) $URL);
@@ -521,7 +520,7 @@ class BackendBlogModel
 		$db = BackendModel::getDB();
 
 		// new item
-		if($itemId === null)
+		if($id === null)
 		{
 			// get number of categories with this URL
 			$number = (int) $db->getVar('SELECT COUNT(i.id)
@@ -549,7 +548,7 @@ class BackendBlogModel
 											FROM blog_posts AS i
 											INNER JOIN meta AS m ON i.meta_id = m.id
 											WHERE i.language = ? AND m.url = ? AND i.id != ?',
-											array(BL::getWorkingLanguage(), $URL, $itemId));
+											array(BL::getWorkingLanguage(), $URL, $id));
 
 			// already exists
 			if($number != 0)
@@ -558,7 +557,7 @@ class BackendBlogModel
 				$URL = BackendModel::addNumber($URL);
 
 				// try again
-				return self::getURL($URL, $itemId);
+				return self::getURL($URL, $id);
 			}
 		}
 
@@ -628,10 +627,10 @@ class BackendBlogModel
 
 
 	/**
-	 * Inserts a blogpost into the database
+	 * Inserts an item into the database
 	 *
 	 * @return	int
-	 * @param	array $item						The data to insert.
+	 * @param	array $item		The data to insert.
 	 */
 	public static function insert(array $item)
 	{
@@ -650,7 +649,7 @@ class BackendBlogModel
 	 * Inserts a new category into the database
 	 *
 	 * @return	int
-	 * @param	array $item						The data for the category to insert.
+	 * @param	array $item		The data for the category to insert.
 	 */
 	public static function insertCategory(array $item)
 	{
@@ -669,15 +668,12 @@ class BackendBlogModel
 	 * Recalculate the commentcount
 	 *
 	 * @return	bool
-	 * @param	array $ids						The id(s) of the post wherefor the commentcount should be recalculated.
+	 * @param	array $ids	The id(s) of the post wherefor the commentcount should be recalculated.
 	 */
-	public static function reCalculateCommentCount($ids)
+	public static function reCalculateCommentCount(array $ids)
 	{
-		// make sure $ids is an array
-		$ids = (array) $ids;
-
 		// validate
-		if(!$ids) return false;
+		if(empty($ids)) return false;
 
 		// make unique ids
 		$ids = array_unique($ids);
@@ -693,7 +689,7 @@ class BackendBlogModel
 												GROUP BY i.post_id',
 												array('published', BL::getWorkingLanguage(), 'active'));
 
-		// loop posts
+		// loop items
 		foreach($ids as $id)
 		{
 			// get count
@@ -708,10 +704,10 @@ class BackendBlogModel
 
 
 	/**
-	 * Update an existing blogpost
+	 * Update an existing item
 	 *
 	 * @return	int
-	 * @param	array $item						The new data.
+	 * @param	array $item		The new data.
 	 */
 	public static function update(array $item)
 	{
@@ -763,7 +759,7 @@ class BackendBlogModel
 	 * Update an existing category
 	 *
 	 * @return	int
-	 * @param	array $item						The new data.
+	 * @param	array $item		The new data.
 	 */
 	public static function updateCategory(array $item)
 	{
@@ -782,7 +778,7 @@ class BackendBlogModel
 	 * Update an existing comment
 	 *
 	 * @return	int
-	 * @param	array $item			The new data.
+	 * @param	array $item		The new data.
 	 */
 	public static function updateComment(array $item)
 	{
@@ -795,16 +791,16 @@ class BackendBlogModel
 	 * Updates one or more comments' status
 	 *
 	 * @return	void
-	 * @param	array $ids						The id(s) of the comment(s) to change the status for.
-	 * @param	string $status					The new status.
+	 * @param	array $ids			The id(s) of the comment(s) to change the status for.
+	 * @param	string $status		The new status.
 	 */
 	public static function updateCommentStatuses($ids, $status)
 	{
 		// make sure $ids is an array
 		$ids = (array) $ids;
 
-		// get blogpost ids
-		$postIds = (array) BackendModel::getDB()->getColumn('SELECT i.post_id
+		// get ids
+		$itemIds = (array) BackendModel::getDB()->getColumn('SELECT i.post_id
 																FROM blog_comments AS i
 																WHERE i.id IN ('. implode(',', $ids) .')');
 
@@ -815,7 +811,7 @@ class BackendBlogModel
 											array((string) $status));
 
 		// recalculate the comment count
-		if(!empty($postIds)) self::reCalculateCommentCount($postIds);
+		if(!empty($itemIds)) self::reCalculateCommentCount($itemIds);
 
 		// invalidate the cache for blog
 		BackendModel::invalidateFrontendCache('blog', BL::getWorkingLanguage());