+++ default_www/backend/modules/pages/engine/model.php
@@ -358,7 +358,7 @@ class BackendPagesModel
 	 * @param	array $template			The template data.
 	 * @param	bool[optional] $large	Will the HTML be used in a large version?
 	 */
-	public static function buildTemplateHTML($template, $large = false)
+	private static function buildTemplateHTML($template, $large = false)
 	{
 		// validate
 		if(!isset($template['data']['format'])) throw new BackendException('Invalid template-format.');
@@ -909,10 +909,10 @@ class BackendPagesModel
 		// get child
 		$childId = (int) BackendModel::getDB()->getVar('SELECT i.id
 														FROM pages AS i
-														WHERE i.parent_id = ? AND i.status = ?
+														WHERE i.parent_id = ? AND i.status = ? AND i.language = ?
 														ORDER BY i.sequence ASC
 														LIMIT 1',
-														array($pageId, 'active'));
+														array($pageId, 'active', BL::getWorkingLanguage()));
 
 		// return
 		if($childId != 0) return $childId;
@@ -1168,6 +1168,9 @@ class BackendPagesModel
 			$row['data'] = unserialize($row['data']);
 			$row['has_block'] = false;
 
+			// reset
+			if(isset($row['data']['default_extras_'. BL::getWorkingLanguage()])) $row['data']['default_extras'] = $row['data']['default_extras_'. BL::getWorkingLanguage()];
+
 			// any extras?
 			if(isset($row['data']['default_extras']))
 			{
@@ -1175,7 +1178,7 @@ class BackendPagesModel
 				foreach($row['data']['default_extras'] as $value)
 				{
 					// store if the module has blocks
-					if(SpoonFilter::isInteger($value) && isset($extras[$value]) && $extras[$value]['type']) $row['has_block'] = true;
+					if(SpoonFilter::isInteger($value) && isset($extras[$value]) && $extras[$value]['type'] == 'block') $row['has_block'] = true;
 				}
 			}
 
@@ -1395,7 +1398,10 @@ class BackendPagesModel
 	{
 		// redefine
 		$URL = (string) $URL;
-		$parentId = (int) $parentId;
+		$parentIds = array((int) $parentId);
+
+		// 0, 1, 2, 3, 4 are all toplevels, so we should place them on the same level
+		if($parentId == 0 || $parentId == 1 || $parentId == 2 || $parentId == 3 || $parentId == 4) $parentIds = array(0, 1, 2, 3, 4);
 
 		// get db
 		$db = BackendModel::getDB();
@@ -1407,8 +1413,8 @@ class BackendPagesModel
 			$number = (int) $db->getVar('SELECT COUNT(i.id)
 										FROM pages AS i
 										INNER JOIN meta AS m ON i.meta_id = m.id
-										WHERE i.parent_id = ? AND i.status = ? AND m.url = ? AND i.language = ?',
-										array($parentId, 'active', $URL, BL::getWorkingLanguage()));
+										WHERE i.parent_id IN('. implode(',', $parentIds) .') AND i.status = ? AND m.url = ? AND i.language = ?',
+										array('active', $URL, BL::getWorkingLanguage()));
 
 			// no items?
 			if($number != 0)
@@ -1428,8 +1434,8 @@ class BackendPagesModel
 			$number = (int) $db->getVar('SELECT COUNT(i.id)
 										FROM pages AS i
 										INNER JOIN meta AS m ON i.meta_id = m.id
-										WHERE i.parent_id = ? AND i.status = ? AND m.url = ? AND i.id != ? AND i.language = ?',
-										array($parentId, 'active', $URL, $id, BL::getWorkingLanguage()));
+										WHERE i.parent_id IN('. implode(',', $parentIds) .') AND i.status = ? AND m.url = ? AND i.id != ? AND i.language = ?',
+										array('active', $URL, $id, BL::getWorkingLanguage()));
 
 			// there are items so, call this method again.
 			if($number != 0)