+++ default_www/frontend/core/engine/block.php
@@ -142,7 +142,7 @@ class FrontendBlockExtra extends FrontendBaseObject
 				foreach($this->config->getPossibleActions() as $actionName)
 				{
 					// get action that should be passed as parameter
-					$actionURL = FrontendLanguage::getAction(SpoonFilter::toCamelCase($actionName));
+					$actionURL = FL::act(SpoonFilter::toCamelCase($actionName));
 
 					// the action is the requested one
 					if($actionURL == $actionParameter)