+++ default_www/frontend/core/engine/frontend.php
@@ -19,6 +19,9 @@ class Frontend
 	 */
 	public function __construct()
 	{
+		// initialize Facebook
+		$this->initializeFacebook();
+
 		// create URL-object
 		new FrontendURL();
 
@@ -27,6 +30,39 @@ class Frontend
 
 		// create and set page reference
 		new FrontendPage();
+	}
+
+
+	/**
+	 * Initialize Facebook
+	 *
+	 * @return	void
+	 */
+	private function initializeFacebook()
+	{
+		// get settings
+		$facebookApplicationId = FrontendModel::getModuleSetting('core', 'facebook_app_id');
+		$facebookApplicationSecret = FrontendModel::getModuleSetting('core', 'facebook_app_secret');
+		$facebookApiKey = FrontendModel::getModuleSetting('core', 'facebook_api_key');
+
+		// needed data available?
+		if($facebookApplicationId != '' && $facebookApplicationSecret != '' && $facebookApiKey != '')
+		{
+			// require
+			require_once 'external/facebook.php';
+
+			// create instance
+			$facebook = new Facebook($facebookApiKey, $facebookApplicationSecret, $facebookApplicationId);
+
+			// get the cookie
+			$data = $facebook->getCookie();
+
+			// set the token if available
+			if(isset($data['access_token'])) $facebook->setToken($data['access_token']);
+
+			// store in reference
+			Spoon::setObjectReference('facebook', $facebook);
+		}
 	}
 }