+++ default_www/frontend/init.php
@@ -276,6 +276,7 @@ class FrontendInit
 	public static function exceptionHandler($exception, $output)
 	{
 		// redefine
+		$exception = $exception;
 		$output = (string) $output;
 
 		// mail it?
@@ -290,7 +291,7 @@ class FrontendInit
 			$headers .= "From: Spoon Library <no-reply@spoon-library.com>\n";
 
 			// send email
-			@mail(SPOON_DEBUG_EMAIL, 'Exception Occured', $output, $headers);
+			@mail(SPOON_DEBUG_EMAIL, 'Exception Occured ('. SITE_DOMAIN .')', $output, $headers);
 		}
 
 		// build HTML for nice error