+++ default_www/frontend/modules/blog/actions/detail.php
@@ -139,7 +139,7 @@ class FrontendBlogDetail extends FrontendBaseBlock
 	{
 		// create form
 		$this->frm = new FrontendForm('comment');
-		$this->frm->setAction($this->frm->getAction() .'#'. FL::getAction('Comment'));
+		$this->frm->setAction($this->frm->getAction() .'#'. FL::act('Comment'));
 
 		// init vars
 		$author = (SpoonCookie::exists('comment_author')) ? SpoonCookie::get('comment_author') : null;
@@ -172,7 +172,7 @@ class FrontendBlogDetail extends FrontendBaseBlock
 		$rssCommentsLink = FrontendNavigation::getURLForBlock('blog', 'article_comments_rss') .'/'. $this->record['url'];
 
 		// add RSS-feed into the metaCustom
-		$this->header->addMetaCustom('<link rel="alternate" type="application/rss+xml" title="'. vsprintf(FL::getMessage('CommentsOn'), array($this->record['title'])) .'" href="'. $rssCommentsLink .'" />');
+		$this->header->addMetaCustom('<link rel="alternate" type="application/rss+xml" title="'. vsprintf(FL::msg('CommentsOn'), array($this->record['title'])) .'" href="'. $rssCommentsLink .'" />');
 
 		// build Facebook Open Graph-data
 		if(FrontendModel::getModuleSetting('core', 'facebook_admin_ids', null) !== null)
@@ -213,17 +213,14 @@ class FrontendBlogDetail extends FrontendBaseBlock
 		$this->header->setMetaKeywords($this->record['meta_keywords'], ($this->record['meta_keywords_overwrite'] == 'Y'));
 
 		// assign article
-		$this->tpl->assign('blogArticle', $this->record);
-
-		// assign article tags
-		$this->tpl->assign('blogArticleTags', $this->record['tags']);
+		$this->tpl->assign('item', $this->record);
 
 		// count comments
 		$commentCount = count($this->comments);
 
 		// assign the comments
-		$this->tpl->assign('blogCommentsCount', $commentCount);
-		$this->tpl->assign('blogComments', $this->comments);
+		$this->tpl->assign('commentsCount', $commentCount);
+		$this->tpl->assign('comments', $this->comments);
 
 		// options
 		if($commentCount > 1) $this->tpl->assign('blogCommentsMultiple', true);
@@ -237,10 +234,10 @@ class FrontendBlogDetail extends FrontendBaseBlock
 		if($this->URL->getParameter('comment', 'string') == 'true') $this->tpl->assign('commentIsAdded', true);
 
 		// assign settings
-		$this->tpl->assign('blogSettings', $this->settings);
+		$this->tpl->assign('settings', $this->settings);
 
 		// assign navigation
-		$this->tpl->assign('blogNavigation', FrontendBlogModel::getNavigation($this->record['id']));
+		$this->tpl->assign('navigation', FrontendBlogModel::getNavigation($this->record['id']));
 	}
 
 
@@ -333,14 +330,14 @@ class FrontendBlogDetail extends FrontendBaseBlock
 				// append a parameter to the URL so we can show moderation
 				if(strpos($redirectLink, '?') === false)
 				{
-					if($comment['status'] == 'moderation') $redirectLink .= '?comment=moderation#'.FL::getAction('Comment');
-					if($comment['status'] == 'spam') $redirectLink .= '?comment=spam#'.FL::getAction('Comment');
+					if($comment['status'] == 'moderation') $redirectLink .= '?comment=moderation#'.FL::act('Comment');
+					if($comment['status'] == 'spam') $redirectLink .= '?comment=spam#'.FL::act('Comment');
 					if($comment['status'] == 'published') $redirectLink .= '?comment=true#comment-'. $comment['id'];
 				}
 				else
 				{
-					if($comment['status'] == 'moderation') $redirectLink .= '&comment=moderation#'.FL::getAction('Comment');
-					if($comment['status'] == 'spam') $redirectLink .= '&comment=spam#'.FL::getAction('Comment');
+					if($comment['status'] == 'moderation') $redirectLink .= '&comment=moderation#'.FL::act('Comment');
+					if($comment['status'] == 'spam') $redirectLink .= '&comment=spam#'.FL::act('Comment');
 					if($comment['status'] == 'published') $redirectLink .= '&comment=true#comment-'. $comment['id'];
 				}